requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
attr==0.3.1
certifi==2022.5.18.1; python_version >= '3.6'
charset-normalizer==2.0.12; python_version >= '3'
dj-database-url==0.5.0
django-cors-headers==3.12.0
django-filter==21.1
django-heroku==0.3.1
django==4.0.4
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
googlemaps==4.6.0
gunicorn==20.1.0
idna==3.3; python_version >= '3'
importlib-metadata==4.11.4; python_version < '3.10'
markdown==3.3.7
pillow==9.1.1
psycopg2==2.9.3
pyjwt==2.4.0; python_version >= '3.6'
pytz==2022.1
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==62.3.2; python_version >= '3.7'
spatialite==0.0.3
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2022.1; sys_platform == 'win32'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uuid==1.30
whitenoise==6.1.0
zipp==3.8.0; python_version >= '3.7'
